[CDEC-328] Implement `deriveIndexedBi` TH function
`deriveIndexedBi` generates `Bi` encoders and decoders for a given
datatype. It is similar to `deriveSimpleBi` but doesn't require named
fields. Instead, it uses zero-indexed integers to label the fields.

As with `deriveSimpleBi`, the fields may be permuted in the declaration
which results in permutation of binary data in the encoding.